pyproject.toml
[tool.poetry]
name = "cas-visualizer"
version = "0.2.0"
license = "MIT"
description = "Visualize CAS (Common Annotation System) annotations in multiple formats: spaCy-style HTML spans, dependency trees, tables, heatmaps, and DOCX with colored annotations."
authors = [
    "Torsten Zesch <torsten.zesch@gmail.com>",
    "Dominic Gardner <dmnc.crl.grdnr@gmail.com>",
]
readme = "README.md"
packages = [{include = "cas_visualizer"}]
repository = "https://github.com/zesch/cas-visualizer"
keywords = ["cas", "visualization", "annotation", "nlp", "spacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pandas = ">=1.3.0"
dkpro-cassis = ">=0.10.0"
spacy = ">=3.2.0"
python-docx = "^1.2.0"
udapi = "^0.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "^1.5.0"
black = "^23.9.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
sphinx = "^7.2.0"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ============= Tool Configurations =============

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
branch = true
source = ["cas_visualizer"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
precision = 2
show_missing = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true
plugins = []

[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*", "*/settings/*"]
known_first_party = ["cas_visualizer"]
known_third_party = ["cassis", "pandas", "spacy", "udapi"]
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]  # Black compatibility
exclude = [".git", "__pycache__", ".tox", ".venv", "build", "dist"]

[tool.pylint]
max-line-length = 100
disable = ["C0111", "R0903", "R0913"]
